Final submit

+++ src/main/java/com/example/resturant_management_system/model/response/StaffInfoResponse.java
@@ -15,5 +15,5 @@ public class StaffInfoResponse {
     private Double basicSalary;
     private String nextShift; // "2025-01-26, 09:00 AM - 05:00 PM"
     private String attendanceOverview; // "85% Present"
-    private long totalHoursWorked; // Total hours
+    private double totalHoursWorked; // Total hours
 }
